+++ packages/moonstone/VirtualList/GridListImageItem.js
@@ -35,7 +35,7 @@ const defaultPlaceholder =
  * @public
  * @deprecated since 1.14.0. Replaced by [moonstone/GridListImageItem.GridListImageItemBase]{@link moonstone/GridListImageItem.GridListImageItemBase}
  */
-const GridListImageItemBase = kind({
+const GridListImageItemBase = deprecate(kind({
 	name: 'GridListImageItem',
 
 	propTypes: /** @lends moonstone/VirtualList.GridListImageItemBase.prototype */ {
@@ -170,7 +170,7 @@ const GridListImageItemBase = kind({
 			</div>
 		);
 	}
-});
+}), {name: 'moonstone/VirtualList.GridListImageItemBase', since: '1.14.0', until: '2.0.0', replacedBy: 'moonstone/GridListImageItemBase'});
 
 /**
  * {@link moonstone/VirtualList.GridListImageItem} is a GridListImageItem with
@@ -194,10 +194,13 @@ const GridListImageItem = MarqueeController(
 	{marqueeOnFocus: true},
 	Spottable(
 		Skinnable(
-			GridListImageItemBase
+			deprecate(
+				GridListImageItemBase,
+				{name: 'moonstone/VirtualList.GridListImageItem', since: '1.14.0', until: '2.0.0', replacedBy: 'moonstone/GridListImageItem'}
+			)
 		)
 	)
 );
 
-export default deprecate(GridListImageItem, {name: 'moonstone/VirtualList.GridListImageItem', since: '1.14.0', until: '2.0.0', replacedBy: 'moonstone/GridListImageItem'});
+export default GridListImageItem;
 export {GridListImageItem, GridListImageItemBase};